Create README.md
This was a practical activity undertaken as part of the Cisco DevNet Associate qualification. It is a piece of Python code illustrating how to use multiple API calls
to automate exisiting Cisco platforms in this case Webex. The code creates a message within a Webex room which prints the name of a city to the message board of the 
room. Based on this information the code takes the name and calls the Mapquest API to get the GPS co-ordinates of the city in question and once it has the
co-ordinates makes a further  API call to the ISS Pass times API which returns the pass times of the International Space Station as it passes overhead
the mentioned city.